WIP: define bottom-up induction principle
There seems to be a bug where the Qed check at a Program Fixpoint
doesn't terminate even though there are no goals left. I'm going to
let Coq compile overnight to see if it manages. To that end I've
separated the code into two files: one for trees and one for terms.